[sourcekitd] Rename module SwiftSourceKit to SwiftLang. (#14497)
The sourcekitd client library provides parsing APIs for SwiftSyntax users.
The internal use of sourcekit service is an implementation detail that end users
shouldn't worry about.
