+++ test/SwiftSyntax/DeserializeFile.swift
@@ -6,7 +6,7 @@
 import StdlibUnittest
 import Foundation
 import SwiftSyntax
-import SwiftSourceKit
+import SwiftLang
 
 func getInput(_ file: String) -> URL {
   var result = URL(fileURLWithPath: #file)
@@ -20,7 +20,7 @@ var DecodeTests = TestSuite("DecodeSyntax")
 
 DecodeTests.test("Basic") {
   expectDoesNotThrow({
-    let content = try SourceFileSyntax.encodeSourceFileSyntax(getInput("visitor.swift"))
+    let content = try SwiftLang.parse(getInput("visitor.swift"))
     let source = try String(contentsOf: getInput("visitor.swift"))
     let parsed = try SourceFileSyntax.decodeSourceFileSyntax(content)
     expectEqual("\(parsed)", source)
