+++ tools/SwiftSourceKitClient/CMakeLists.txt
@@ -2,7 +2,8 @@ set(EXTRA_COMPILE_FLAGS "-F" "${SWIFT_LIBRARY_OUTPUT_INTDIR}")
 set(EXTRA_LINKER_FLAGS "-Xlinker" "-rpath" "-Xlinker" "${SWIFT_LIBRARY_OUTPUT_INTDIR}"
     "-Xlinker" "-F" "-Xlinker" "${SWIFT_LIBRARY_OUTPUT_INTDIR}")
 
-add_swift_library(swiftSwiftSourceKit SHARED
+add_swift_library(swiftSwiftLang SHARED
+  SwiftLang.swift
   SourceKitdClient.swift
   SourceKitdRequest.swift
   SourceKitdResponse.swift
